requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
google-auth-httplib2==0.2.0
gdown==5.1.0
requests==2.31.0
